Initial commit
Somewhat-clean version of the prototype. Some noticable artifacts are present in the output, documentation is practically nonexistent and readme.md spends all its time on justifying the project's very existence, and with no scrolling or any indication how much time the presentation will take, this whole dead log of a software may very well be considered to be an unusable waste of time to everyone involved.